requirements.txt
discord.py
requests
nltk
emojis
pyyaml
